+++ src/collective/volto/formsupport/restapi/services/submit_form/post.py
@@ -259,6 +259,14 @@ class SubmitPost(Service):
         msg["To"] = mto
         msg["Reply-To"] = mreply_to
 
+        msg.replace_header("Content-Type", 'text/html; charset="utf-8"')
+
+        headers_to_forward = self.block.get("httpHeaders", [])
+        for header in headers_to_forward:
+            header_value = self.request.get(header)
+            if header_value:
+                msg[header] = header_value
+
         self.manage_attachments(msg=msg)
         self.send_mail(msg=msg, charset=charset)